Fix potential NPE in TableDrop
I suspect that this will not get through a working MySQL server, but if we've just tested the database for being null, calling a method on it is quite dangerous.

This code will still return "newSchema" if d == null && ifExists, but will use the existing return.

Incidentally, is there any difference between returning originalSchema or newSchema in this case? Should we return originalSchema if nothing's changed?