support CREATE TABLE SELECT * FROM `foo`
with row-based replication this gets transformed to:

BEGIN
CREATE table t ( ... )
ROWS
COMMIT

so we have to support CREATE TABLE ddl inside a transaction.  Note that
this means we could process the CREATE TABLE, flush the binlog position,
crash, and then restart inside a transaction, missing the BEGIN event.
For this reason we downgrade the started-replication-outside-transaction
case to a warning.