handle the annoyingly large vocabulary of CHECK()
the problem:
mysql allows CHECK( statement ) constraints that have full sql
statement capabilities inside the CHECK( .. ).  These statements do not
a damn thing, but we parse them nonetheless.

the solution:
Our parser is good enough to recognize parenthesis and then a bunch of
arbitrary tokens.  When we see that we've hit an unparsable section, we
drop out of the parser and snip out matching parenthesis from the token
stream, and then re-parse.