refactor ColumnDef classes
In preparation for serializing ColumnDef classes, we move them to a more
"proper" hierarchy in which only string columns have a character set,
only enum columns have enum values, and only integer columns have a
signed flag.

There's still some type-checking hacks that it'd be good to remove by
delegating the methods to the objects, but this is good for now.