Make schema storage highly more efficient
Instead of storing a giant snapshot with each
schema change, we now keep a base snapshot and
a list of alters.  The alters are serialized
as JSON.

In order to do this, I've had to remove the 
functionality that cleans up old schemas.  At a 
certain point, it may be better to re-save
a snapshot (maybe after like 100k schema alters).
Will do that in a later PR