setup some saner kafka defaults, rescue MessageTooLarge
handle the fact that we've dropped a too-large message, and set retries
to "1" by default -- the default of "0" meant that we could drop
messages when a new leader was -elected, which seems pretty insane to
me.  Also set retries to 1, as otherwise it's pretty easy to drop
messages on the floor when a new leader is elected.