Do some work on Maxwell memory pressure
Under load in production, it was fairly easy to get Maxwell to run
itself out of memory, especially with a transaction that involves many
>10mb rows.

This changes the way that we flush our list of transactions to disk;
previously we had kept a fixed size nubmer of RowMap objects in memory,
now we make an educated guess about the total size of the in-memory
cache and don't let it grow larger than 25% of heap memory.

There's also a change to the Kafka producer to allow it to release its
keys and values sooner.

With this change, a set of updates that previously required a 4gb heap
to consume now can be safely consumed with Maxwell running at -Xmx300m.