cleanup maxwell schema-database interfaces
There's a need at boot-up time to have access to a mysql connection (to
the schema-store) with no database set, so that we can create it.  This
was permeating all the interfaces, causing us to pass
`config.databaseName` all around the application.  This moves the
pre-database creation usages out of the main code and into a side path.