separate KafkaProducer#push from the main thread
This should be worth roughly 25% of cpu-concurrency, by running
binlog-event-to-RowMap code on a different thread than the
json-marshalling and kafka-data-compression/batching.  We could go
deeper than this by also parceling out the json-marshal calls into a
separate thread, but at that point I think I'd adopt rx-java or some
kind of pipeline framework.