Remove redundant log.
If MySQL service is not running, maxwell when started will abort due to
Communication link failure.  However we see the logs printed twice.

➜  maxwell git:(master) ✗ bin/maxwell --user='maxwell' --password='maxwell' --host='127.0.0.1' --producer=kafka --kafka.bootstrap.servers=localhost:9092 --kafka_topic=test
Using kafka version: 0.11.0.1
11:29:13,725 WARN  MaxwellMetrics - Metrics will not be exposed: metricsReportingType not configured.
11:29:14,114 ERROR Maxwell - SQLException: Communications link failure

The last packet sent successfully to the server was 0 milliseconds ago. The driver has not received any packets from the server.
11:29:14,114 ERROR Maxwell - Communications link failure

The last packet sent successfully to the server was 0 milliseconds ago. The driver has not received any packets from the server.