improved configuration validation
issue #952 showed that it was pretty easy to get embedded-maxwell into a
bad state with a missing MaxwellFilter class, so I'm moving the call to
MaxwellConfig#validate out of the options parser and into the main
startup line.

I've also re-coded some of the partition-by validation code to apply to
other producers and the new partitionBy config keys.