add maxwell-benchmark
this is a bit of code that boots up a server, generates data, and then
leaves that data in place for you to run performance tests upon.
For now we use a custom producer to know when to shutdown, but in the
long run it might be useful to be able to benchmark other producers,
too.