re-do stuck head detection
We now ask the question, "how many messages have gone by while the head
of the queue has been stuck"?

Note that all of these are mildly dangerous assumptions; the head of the
queue *could* be a single kafka partition offline, while the other
partitions are succeeding.  In this case, we're not doing anything
harmful by waiting for the partition to come back.