issue #1216, deal with bizarre ordering issues on column renames
there's some implementation detail in mysql that allows for a column renamed
later in a DDL statement to be referenced by the new name (and, in fact,
requires it).  We've always processed things in a single shot,
sequentially, but this breaks that assumption.  So we special case this
thing and just backtrack and patch it up at the end