refactor: tease Replicator out of Bootstrapper
Previously bootstrapper would ask the Replicator for a current copy of
the schema.  This was actually race-condition prone and also not great
for the goal of running bootstraps independent of rows coming through
the replication pipeline.

Instead, we capture the current copy of the schema ala minute.  There's
a bad feature of outputting the "current" schema-id in the bootstrap
that we'll have to keep, inaccurate though it may be.