:bug: use meter for counts instead of Counter
dropwizard metrics Meter is mapped to statsd's Counter.
IF we use dropwizard Counter then the metric will endup as a gauge in Datadog.
Reference https://github.com/coursera/metrics-datadog/issues/45

The referenced issue also mentioned a version downgrade for metrics-datadog. I tried it locally and the version didn't seem to make a difference.