fixes, sort of, for some of #1474.
this half-hearted change fixes our processing of dates with a zero-year.
There doesn't seem to be a great way to represent these properly in
java-land, as they're not really valid dates, at least not in the
gregorian calendar.  Maybe the proper solution is to just have a mode of
the binlog-connector where it apes the mysql display function and
returns strings.