Make all debug statements parameterized
Most of the logging in Maxwell is currently done by concatenating strings, however one of the main benefits of using SLF4j is that you can skip the cost of concatenating log message strings and calling the toString method on parameters when you don't log something. This is particularly relevant for debug logging because it is usually turned off. The same should be done for all other logging calls for consistency, but those changes are not part of this PR.

There are some places where there is both a isDebugEnabled() check and parameterization, I added isDebugEnabled where it was relevant (like calls to an alternative toString method for debugging purposes) and left it alone when it was already there. Whiparameterization is  redundant in these cases I think consistency of using parameterization is more important.

Reference: http://www.slf4j.org/faq.html#logging_performance