Add completion logging to Schema Compactor
In environments with very many databases and tables, it can take a long time (1hr+) to compact the schema changes. However, it is difficult to figure this out unless you are watching the processlist as there are no logs or metrics to indicate completion nor any indication in the `maxwell` database of how long it took.

This PR adds two additional log lines to MysqlSchemaCompactor to indicate completion of the commit and slow deletion phases.
For most users, they will never see the logs because they rarely or never compact.